# optimize DELTA_LENGTH_BYTE_ARRAY encoding