--- a/encoding/delta/length_byte_array_amd64.go
+++ b/encoding/delta/length_byte_array_amd64.go
@@ -3,4 +3,7 @@
 package delta
 
 //go:noescape
+func encodeByteArrayLengths(lengths []int32, offsets []uint32)
+
+//go:noescape
 func decodeByteArrayLengths(offsets []uint32, lengths []int32) (lastOffset uint32, invalidLength int32)