Follow up to https://github.com/segmentio/parquet-go/pull/289, this PR optimizes the encoding of DELTA_LENGTH_BYTE_ARRAY pages.

```
name                                       old time/op    new time/op    delta
Encode/DELTA_LENGTH_BYTE_ARRAY/byte_array    18.2µs ± 0%    13.7µs ± 0%  -24.95%  (p=0.000 n=10+10)

name                                       old speed      new speed      delta
Encode/DELTA_LENGTH_BYTE_ARRAY/byte_array  5.97GB/s ± 0%  7.96GB/s ± 0%  +33.24%  (p=0.000 n=10+10)

name                                       old value/s    new value/s    delta
Encode/DELTA_LENGTH_BYTE_ARRAY/byte_array      549M ± 0%      731M ± 0%  +33.24%  (p=0.000 n=10+10)
```